Tried simplifying type conversions among float, half, and bfloat16

+++ include/flashinfer/attention/cascade.cuh
@@ -380,7 +380,8 @@ __global__ void PersistentVariableLengthMergeStatesKernel(
     if (num_index_sets == 0) {
       vec_t<DTypeO, vec_size> v;
 #ifdef __HIPCC__
-      v.fill(DTypeO(__float2half(0.f)));
+      // v.fill(DTypeO(__float2half(0.f)));
+      v.fill(DTypeO(__float2bfloat16(0.f)));
 #else
       v.fill(DTypeO(0.f));
 #endif
